+++ POS/POS.Logic/PayoutSettings/PayoutEventType.cs
@@ -7,6 +7,8 @@
         CashAdded = 2002,
         CashRemoved = 2003,
         VoucherSuccessfullyPaid = 2004,
-        ReprintPayoutReceipt = 2005
+        ReprintPayoutReceipt = 2005,
+        PrintPayoutReceipt = 2006,
+        CashDrawerOpened = 2007
     }
 }